Safari-611.1.10/LayoutTests/js/dom/modules/import-meta-url-expected.txt

Test import.meta.url.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS url.endsWith("js/dom/modules/import-meta-url.html") is true
PASS url.endsWith("js/dom/modules/script-tests/import-meta-url-top-level.js") is true
PASS url.endsWith("js/dom/modules/script-tests/import-meta-url-second-level.js") is true
PASS successfullyParsed is true

TEST COMPLETE
